hm: 'stream_length' contains the original stream size.
Cause the huffman coding has a bit granularity, most of the time the
encoded file has some garbage zero to the end that have not to be
decoded; one possibility is to create a code for the end of stream but
is cubersome to do, another option is to add to the encoded file header
the size of the original stream.

WARNING: only on machine with the same endianess this works.